From long to int

+++ timectxsw.c
@@ -44,9 +44,9 @@ static inline long long unsigned time_ns(struct timespec* const ts) {
 int main(int argc, char **argv) {
   const int iterations = 500000;
 
-  unsigned long *results = malloc(sizeof(unsigned long) * iterations);
-  memset(results, 0, sizeof(unsigned long) * iterations);
-  int ret = mlock(results, sizeof(unsigned long) * iterations);
+  unsigned long *results = malloc(sizeof(unsigned int) * iterations);
+  memset(results, 0, sizeof(unsigned int) * iterations);
+  int ret = mlock(results, sizeof(unsigned int) * iterations);
 
   const int shm_id = shmget(IPC_PRIVATE, sizeof (int), IPC_CREAT | 0666);
   const pid_t other = fork();
@@ -69,7 +69,7 @@ int main(int argc, char **argv) {
         sched_yield();
       }
       stop = rdtsc();
-      results[i] = (unsigned long)(stop-start);
+      results[i] = (unsigned int)(stop-start);
       start = stop;
     }
     return 0;
@@ -89,7 +89,7 @@ int main(int argc, char **argv) {
       sched_yield();
     }
     stop = rdtsc();
-    results[i] = (unsigned long)(stop-start);
+    results[i] = (unsigned int)(stop-start);
     start = stop;
   }
   const long long unsigned delta = time_ns(&ts) - start_ns;
@@ -101,7 +101,7 @@ int main(int argc, char **argv) {
       return 1;
     }
     fwrite(&delta, sizeof(unsigned long long), 1, out);
-    fwrite(results, sizeof(unsigned long), iterations, out);
+    fwrite(results, sizeof(unsigned int), iterations, out);
     fclose(out);
   }
